Reset RSpec configuration
https://www.pivotaltracker.com/story/show/100367828

The old tests dated back to the very start of the project and had not
been kept in line with the changes to the app. When running the suite
as a whole, over 60% of the tests were failing. It was deiced that it
would be easier to start anew rather than trying to fix them
retrospectively.

* Remove old tests and factories.
* Upgrade RSpec.
* Remove Capybara and assocated drivers.
* Add Guard.