Record when a booking is cancelled by an admin user
https://www.pivotaltracker.com/story/show/105917930

Continuing the saga of mismatched booking states, the view expects
a booking's `cancel_date` to be set in order to recognise a booking
as having been cancelled. This is now performed when a booking is
cancelled by an admin user.

The bookings highlighted in the Pivotal story have been manually
updated.